feat: add borders and use SVG for scale-h
